Remove Django USWDS Forms
As-is, the package forces Django to be version 1.x. Considering
that the current stable release of Django is version 2.1.7,
using the Django USWDS Forms package doesn't seem warranted.

If time allows, and if the package is otherwise sufficiently
useful, this may be an opportunity to update that package to be
compatible with Django 2.x.