Distill FAC results by agency
The Federal Audit Clearinghouse (FAC) limits the number of results
that can be downloaded, and it does so in a way that poses several
usability challenges. This piece of the prototype is an effort to
make it easier for agency staff to see which single audits need
their attention, by presenting only the single audits 1) for which
their agency is the cognizant[1] agency (i.e., the agency principally
assigned to resolve any cross-cutting findings) and 2) that have
findings.

This commit:
 * Loads data from a local file
 * Prompts the site visitor to select a federal agency
 * Adds an agency lookup table, to facilitate filtering
 * Filters by cognizant agency
 * Adds a placeholder results page
 * Generates a CSV of the relevant results
 * Offers that CSV for download

[1] For further details about what a cognizant agency is, see the
glossary of CFR 200. (Really! It's quite helpful.)