Revert "Add Jupyter for local use, update dependencies"
This reverts commit f936756caa968682aeca08c57c837b87ddf699ea.

A vulnerability has been found in one of Jupyter's dependencies,
Parso. (For details, see https://nvd.nist.gov/vuln/detail/CVE-2019-12760.)

No patched version is available yet, so let's remove Jupyter
entirely for now. If it's relevant to re-add it in the future, we
can do so in a way that's dev-environment-only; as mentioned in
the initial commit, I'd been using it only locally anyway.