Update Download button to download PDFs instead of CSV
Now, instead of triggering the download of a CSV that's been
reduced to the relevant results, the button triggers a download
of the relevant single audit PDFs directly from the Federal Audit
Clearinghouse.

The rationale here is that though the CSV is useful for getting
an overview, the associated PDFs are what contain the actual
findings, which are the more actionable (and necessary) part.

https://github.com/18F/federal-grant-reporting/issues/154