Refactor to download PDFs individually instead of as a ZIP file
This commit restores PDF (single audit package) downloads as well
as waiting for downloads to complete.

Why download the PDFs individually instead of as a ZIP file? Both
because that makes the approach parallel to the one used to
download SF-SAC forms and because the relative granularity will
afford better insight into overall download progress and also fail
more gracefully if the downloads are interrupted.

https://github.com/18F/federal-grant-reporting/issues/148